Use singleton pattern
